Refactor SIRI updaters
